+++ packages/router-core/package.json
@@ -148,6 +148,26 @@
         "default": "./dist/cjs/isServer/client.cjs"
       }
     },
+    "./scroll-restoration-script": {
+      "browser": {
+        "import": {
+          "types": "./dist/esm/scroll-restoration-script/client.d.ts",
+          "default": "./dist/esm/scroll-restoration-script/client.js"
+        },
+        "require": {
+          "types": "./dist/cjs/scroll-restoration-script/client.d.cts",
+          "default": "./dist/cjs/scroll-restoration-script/client.cjs"
+        }
+      },
+      "import": {
+        "types": "./dist/esm/scroll-restoration-script/server.d.ts",
+        "default": "./dist/esm/scroll-restoration-script/server.js"
+      },
+      "require": {
+        "types": "./dist/cjs/scroll-restoration-script/server.d.cts",
+        "default": "./dist/cjs/scroll-restoration-script/server.cjs"
+      }
+    },
     "./package.json": "./package.json"
   },
   "sideEffects": false,
