+++ packages/router-core/src/index.ts
@@ -397,7 +397,6 @@ export { isNotFound, notFound } from './not-found'
 
 export {
   defaultGetScrollRestorationKey,
-  getScrollRestorationScriptForRouter,
   getElementScrollRestorationEntry,
   storageKey,
   scrollRestorationCache,
