+++ packages/router-core/src/scroll-restoration-script/client.ts
@@ -1,0 +1,5 @@
+import type { AnyRouter } from '../router'
+
+export function getScrollRestorationScriptForRouter(_router: AnyRouter) {
+  return null
+}
