+++ packages/router-core/vite.config.ts
@@ -31,6 +31,8 @@ export default mergeConfig(
       './src/index.ts',
       './src/ssr/client.ts',
       './src/ssr/server.ts',
+      './src/scroll-restoration-script/client.ts',
+      './src/scroll-restoration-script/server.ts',
       './src/isServer/server.ts',
       './src/isServer/client.ts',
       './src/isServer/development.ts',
