+++ packages/vue-router/src/scroll-restoration.tsx
@@ -1,5 +1,5 @@
 import * as Vue from 'vue'
-import { getScrollRestorationScriptForRouter } from '@tanstack/router-core'
+import { getScrollRestorationScriptForRouter } from '@tanstack/router-core/scroll-restoration-script'
 import { useRouter } from './useRouter'
 import { ScriptOnce } from './ScriptOnce'
